requirement.txt
streamlit
easyocr
pymupdf
langchain-openai
python-docx
numpy
Pillow
python-dotenv
